Token filter updates
