+++ specification/_types/analysis/kuromoji-plugin.ts
@@ -21,11 +21,17 @@ import { integer } from '@_types/Numeric'
 import { CharFilterBase } from './char_filters'
 import { TokenizerBase } from './tokenizers'
 import { TokenFilterBase } from './token_filters'
+import { StopWords } from './StopWords'
 
 export class KuromojiAnalyzer {
   type: 'kuromoji'
   mode: KuromojiTokenizationMode
   user_dictionary?: string
+}
+
+export class JaStopTokenFilter extends TokenFilterBase {
+  type: 'ja_stop'
+  stopwords?: StopWords
 }
 
 export class KuromojiIterationMarkCharFilter extends CharFilterBase {
